--- Assets/AudioManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f22d77bf532bbea4186bfec0f1cf1c9a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
