+++ Assets/Scripts/PizzaBehaviour.cs
@@ -32,6 +32,7 @@ public class PizzaBehaviour : MonoBehaviour, iPoolerObject
 
     private void Player1Hits()
     {
+        AudioManager.instance.Play("Hit");
         int scoreToAdd = 0;
 
         if (pizzaSize == PizzaSizes.Small)
@@ -57,6 +58,7 @@ public class PizzaBehaviour : MonoBehaviour, iPoolerObject
 
     private void Player2Hits()
     {
+        AudioManager.instance.Play("Hit");
         int scoreToAdd = 0;
 
         if (pizzaSize == PizzaSizes.Small)
