--- Assets/Sound.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 574149de7d1ed5f459a9623945a08781
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
