--- Assets/lockRotation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c676c9d781d84666b4582db6bad592f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
